# Simplify gulmc