--- a/oasislmf/pytools/gul/common.py
+++ b/oasislmf/pytools/gul/common.py
@@ -13,15 +13,6 @@ items_data_type = nb.from_dtype(np.dtype([('item_id', oasis_int),
                                           ('rng_index', oasis_int)
                                           ]))
 
-items_MC_data_type = nb.from_dtype(np.dtype([('item_id', oasis_int),
-                                             ('areaperil_id', areaperil_int),
-                                             ('vulnerability_id', oasis_int),
-                                             ('hazcdf_i', oasis_int),
-                                             ('rng_index', oasis_int),
-                                             ('hazard_rng_index', oasis_int),
-                                             ('eff_vuln_cdf_i', oasis_int),
-                                             ('eff_vuln_cdf_Ndamage_bins', oasis_int)
-                                             ]))
 
 VulnCdfLookup = nb.from_dtype(np.dtype([('start', oasis_int), ('length', oasis_int)]))