--- a/oasislmf/pytools/gulmc/aggregate.py
+++ b/oasislmf/pytools/gulmc/aggregate.py
@@ -38,7 +38,7 @@ def gen_empty_agg_vuln_to_vuln_ids():
 
 
 @njit(cache=True)
-def gen_empty_areaperil_vuln_ids_to_weights():
+def gen_empty_areaperil_vuln_idx_to_weights():
     """Generate empty map to store the weights of individual vulnerability functions in each aggregate vulnerability.
 
     Returns:
@@ -117,12 +117,12 @@ def process_aggregate_vulnerability(aggregate_vulnerability):
     Returns:
         dict[int, list[int]]: map of aggregate vulnerability id to list of vulnerability ids.
     """
-    agg_vuln_to_vuln_id = gen_empty_agg_vuln_to_vuln_ids()
+    agg_vuln_to_vuln_ids = gen_empty_agg_vuln_to_vuln_ids()
 
     if aggregate_vulnerability is not None:
 
         agg_vuln_df = pd.DataFrame(aggregate_vulnerability)
-        # init agg_vuln_to_vuln_id to allow numba to compile later functions
+        # init agg_vuln_to_vuln_ids to allow numba to compile later functions
         # vulnerability_id and aggregate_vulnerability_id are remapped to the internal ids
         # using the vulnd_dict map that contains only the vulnerability_id used in this portfolio.
 
@@ -131,68 +131,28 @@ def process_aggregate_vulnerability(aggregate_vulnerability):
         for agg, grp in agg_vuln_df.groupby('aggregate_vulnerability_id'):
             agg_vuln_id = nb_int32(agg)
 
-            if agg_vuln_id not in agg_vuln_to_vuln_id:
-                agg_vuln_to_vuln_id[agg_vuln_id] = List.empty_list(nb_int32)
+            if agg_vuln_id not in agg_vuln_to_vuln_ids:
+                agg_vuln_to_vuln_ids[agg_vuln_id] = List.empty_list(nb_int32)
 
             for entry in grp['vulnerability_id'].to_list():
-                agg_vuln_to_vuln_id[agg_vuln_id].append(nb_int32(entry))
+                agg_vuln_to_vuln_ids[agg_vuln_id].append(nb_int32(entry))
 
-    return agg_vuln_to_vuln_id
+    return agg_vuln_to_vuln_ids
 
 
 @nb.njit(cache=True)
-def process_vulnerability_weights(areaperil_dict, vuln_dict, aggregate_weights):
-    """Rearrange vulnerability weights from tabular format to a map between (areaperil_id, vulnerability_id) and the vulnerability weight.
-    note: in areaperil_dict, there is no aggregate vulnerability, if one is present in items,
-        the dis-aggregated detailed vuln_in are in areaperil_dict
+def process_vulnerability_weights(areaperil_vuln_i_to_weight, vuln_dict, aggregate_weights):
+    """
+    Polpulate the useful (areaperil_id, vulnerability_i) in areaperil_vuln_i_to_weight with the weight from aggregate_weights
 
     Args:
-        areaperil_dict: dict of areaperil to dict of vuln_id to 0 (0 represent the index but is not set yet)
-        vuln_dict: vuln_dict (Tuple[Dict[int, int]): vulnerability dictionary.
+        areaperil_vuln_i_to_weight: dict of useful (areaperil_id, vulnerability_i) to 0. (weight placeholder to be updated)
+        vuln_dict: vuln_dict (Tuple[Dict[int, int]): vulnerability dictionary, vuln_id => vuln_i.
         aggregate_weights (np.array[VulnerabilityWeight]): vulnerability weights table.
-
-        agg_vuln_to_vuln_id (dict[int, list[int]]): map of aggregate vulnerability id to list of vulnerability ids.
-
-    Returns:
-        dict[AGG_VULN_WEIGHTS_KEY_TYPE, AGG_VULN_WEIGHTS_VAL_TYPE]: map between the areaperil id and the index where the vulnerability function
-          is stored in `vuln_array` and the vulnerability weight.
     """
-    areaperil_vuln_id_to_weight = gen_empty_areaperil_vuln_ids_to_weights()
-
-    if aggregate_weights is None:
-        return areaperil_vuln_id_to_weight
-
-    for areaperil_id in areaperil_dict:
-        for vuln_id in areaperil_dict[areaperil_id]:
-            areaperil_vuln_id_to_weight[(nb_areaperil_int(areaperil_id), vuln_dict[vuln_id])] = nb_oasis_float(0)
-
     for i in range(len(aggregate_weights)):
         rec = aggregate_weights[i]
         if rec['vulnerability_id'] in vuln_dict:
             key = (nb_areaperil_int(rec['areaperil_id']), vuln_dict[rec['vulnerability_id']])
-            if key in areaperil_vuln_id_to_weight:
-                areaperil_vuln_id_to_weight[key] = nb_oasis_float(rec['weight'])
-
-    return areaperil_vuln_id_to_weight
-
-
-@njit(cache=True)
-def map_agg_vuln_ids_to_agg_vuln_idxs(agg_vulns, agg_vuln_to_vuln_id, vuln_dict):
-    """For each aggregate vulnerability listed in `agg_vulns`, map the individual vulnerability_ids that compose it
-    to the indices where they are stored in `vuln_array`.
-
-    Args:
-        agg_vulns (List[int32])
-        agg_vuln_to_vuln_id (dict[int, list[int]]): map of aggregate vulnerability id to list of vulnerability ids.
-        vuln_dict (Tuple[Dict[int, int]): vulnerability dictionary.
-
-    Returns:
-        dict[int, list[int]]: map between aggregate vulnerability id and the list of indices where the individual vulnerability_ids
-          that compose it are stored in `vuln_array`.
-    """
-    agg_vuln_to_vuln_idxs = Dict.empty(nb_int32, List.empty_list(nb_int32))
-
-    for agg in agg_vulns:
-        agg_vuln_to_vuln_idxs[agg] = List([vuln_dict[vuln] for vuln in agg_vuln_to_vuln_id[agg]])
-
-    return agg_vuln_to_vuln_idxs
+            if key in areaperil_vuln_i_to_weight:
+                areaperil_vuln_i_to_weight[key] = nb_oasis_float(rec['weight'])