--- a/oasislmf/pytools/gulmc/common.py
+++ b/oasislmf/pytools/gulmc/common.py
@@ -1,7 +1,7 @@
 import numba as nb
 import numpy as np
 
-from oasislmf.pytools.common.data import areaperil_int, oasis_float, oasis_int, nb_oasis_int
+from oasislmf.pytools.common.data import areaperil_int, oasis_float, oasis_int
 
 # gul stream type
 # probably need to set this dynamically depending on the stream type