--- a/oasislmf/pytools/gulmc/common.py
+++ b/oasislmf/pytools/gulmc/common.py
@@ -98,4 +98,6 @@ NormInversionParameters = nb.from_dtype(np.dtype([('x_min', np.float32),
                                                   ('N', np.int32),
                                                   ('cdf_min', np.float32),
                                                   ('cdf_max', np.float32),
+                                                  ('inv_factor', np.float64),
+                                                  ('norm_factor', np.float64),
                                                   ]))