… tests/assets/test_model_1/expected/exp_res_test_model_1_a1_S1000_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a1_S1000_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a1_S100_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a1_S100_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a1_S10_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a1_S10_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a1_S1_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a1_S1_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a2_S1000_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a2_S1000_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a2_S100_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a2_S100_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a2_S10_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a2_S10_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a2_S1_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a2_S1_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a3_S1000_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a3_S1000_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a3_S100_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a3_S100_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a3_S10_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a3_S10_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a3_S1_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_1/expected/exp_res_test_model_1_a3_S1_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_2/expected/exp_res_test_model_2_a1_S0_L0_ign_corrFalse_rng0_eff_damagFalse.bin
… tests/assets/test_model_2/expected/exp_res_test_model_2_a1_S0_L0_ign_corrFalse_rng0_eff_damagTrue.bin
… tests/assets/test_model_2/expected/exp_res_test_model_2_a1_S0_L0_ign_corrFalse_rng1_eff_damagFalse.bin
… tests/assets/test_model_2/expected/exp_res_test_model_2_a1_S0_L0_ign_corrFalse_rng1_eff_damagTrue.bin
… tests/assets/test_model_2/expected/exp_res_test_model_2_a1_S0_L0_ign_corrTrue_rng0_eff_damagFalse.bin
… tests/assets/test_model_2/expected/exp_res_test_model_2_a1_S0_L0_ign_corrTrue_rng0_eff_damagTrue.bin