@johcarter 2025-04-16
I tested the 'use equi-weight' as default feature for Bangladesh cyclone with
* partially complete weights file
* empty weights file (header only)

The full weights file is all equally weighted so I expected to get the same results for the above cases, and I did (AAL agrees to about 5 s.f)

The only thing that doesn't work is a completely missing weights file which generates a runtime error, but it is reasonable to expect it to be present if aggregate_vulnerability is present.